Refactor MCP tools to improve SIMPLE and COMPOSABLE metrics
- Decomposed `assess.py` (cyclomatic complexity 93 -> 41) into `core.py`, `render.py`, `snapshot.py`, and `worktree.py`.
- Decomposed `evaluate.py` (SIMPLE 0% -> 53%) into `core.py`, `project.py`, and `render.py`.
- Grouped endpoints into sub-packages `assess/` and `evaluate/` to better isolate logic and improve structure.
- Updated `__init__.py` hub and relevant tests to reflect the new paths.
